+++ dsn_test.go
@@ -8,12 +8,14 @@ import (
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/config"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/bind"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/connector"
+	querySql "github.com/ydb-platform/ydb-go-sdk/v3/internal/query/conn"
+	tableSql "github.com/ydb-platform/ydb-go-sdk/v3/internal/table/conn"
 )
 
 func TestParse(t *testing.T) {
-	newConnector := func(opts ...xsql.ConnectorOption) *xsql.Connector {
-		c := &xsql.Connector{}
+	newConnector := func(opts ...connector.Option) *connector.Connector {
+		c := &connector.Connector{}
 		for _, opt := range opts {
 			if opt != nil {
 				if err := opt.Apply(c); err != nil {
@@ -24,6 +26,12 @@ func TestParse(t *testing.T) {
 
 		return c
 	}
+	newTableConn := func(opts ...tableSql.Option) *tableSql.Conn {
+		return tableSql.New(context.Background(), nil, nil, opts...)
+	}
+	newQueryConn := func(opts ...querySql.Option) *querySql.Conn {
+		return querySql.New(context.Background(), nil, nil, opts...)
+	}
 	compareConfigs := func(t *testing.T, lhs, rhs *config.Config) {
 		require.Equal(t, lhs.Secure(), rhs.Secure())
 		require.Equal(t, lhs.Endpoint(), rhs.Endpoint())
@@ -32,22 +40,9 @@ func TestParse(t *testing.T) {
 	for _, tt := range []struct {
 		dsn           string
 		opts          []config.Option
-		connectorOpts []xsql.ConnectorOption
+		connectorOpts []connector.Option
 		err           error
 	}{
-		{
-			dsn: "grpc://localhost:2135/local?go_fake_tx=scripting,scheme",
-			opts: []config.Option{
-				config.WithSecure(false),
-				config.WithEndpoint("localhost:2135"),
-				config.WithDatabase("/local"),
-			},
-			connectorOpts: []xsql.ConnectorOption{
-				xsql.WithFakeTx(xsql.ScriptingQueryMode),
-				xsql.WithFakeTx(xsql.SchemeQueryMode),
-			},
-			err: nil,
-		},
 		{
 			dsn: "grpc://localhost:2135/local",
 			opts: []config.Option{
@@ -75,8 +70,8 @@ func TestParse(t *testing.T) {
 				config.WithEndpoint("localhost:2135"),
 				config.WithDatabase("/local"),
 			},
-			connectorOpts: []xsql.ConnectorOption{
-				xsql.WithDefaultQueryMode(xsql.ScriptingQueryMode),
+			connectorOpts: []connector.Option{
+				connector.WithDefaultQueryMode(tableSql.ScriptingQueryMode),
 			},
 			err: nil,
 		},
@@ -87,9 +82,9 @@ func TestParse(t *testing.T) {
 				config.WithEndpoint("localhost:2135"),
 				config.WithDatabase("/local"),
 			},
-			connectorOpts: []xsql.ConnectorOption{
-				xsql.WithDefaultQueryMode(xsql.ScriptingQueryMode),
-				xsql.WithTablePathPrefix("path/to/tables"),
+			connectorOpts: []connector.Option{
+				connector.WithDefaultQueryMode(tableSql.ScriptingQueryMode),
+				connector.WithQueryBind(bind.TablePathPrefix("path/to/tables")),
 			},
 			err: nil,
 		},
@@ -100,10 +95,10 @@ func TestParse(t *testing.T) {
 				config.WithEndpoint("localhost:2135"),
 				config.WithDatabase("/local"),
 			},
-			connectorOpts: []xsql.ConnectorOption{
-				xsql.WithDefaultQueryMode(xsql.ScriptingQueryMode),
-				xsql.WithTablePathPrefix("path/to/tables"),
-				xsql.WithQueryBind(bind.NumericArgs{}),
+			connectorOpts: []connector.Option{
+				connector.WithDefaultQueryMode(tableSql.ScriptingQueryMode),
+				connector.WithQueryBind(bind.TablePathPrefix("path/to/tables")),
+				connector.WithQueryBind(bind.NumericArgs{}),
 			},
 			err: nil,
 		},
@@ -114,10 +109,10 @@ func TestParse(t *testing.T) {
 				config.WithEndpoint("localhost:2135"),
 				config.WithDatabase("/local"),
 			},
-			connectorOpts: []xsql.ConnectorOption{
-				xsql.WithDefaultQueryMode(xsql.ScriptingQueryMode),
-				xsql.WithTablePathPrefix("path/to/tables"),
-				xsql.WithQueryBind(bind.PositionalArgs{}),
+			connectorOpts: []connector.Option{
+				connector.WithDefaultQueryMode(tableSql.ScriptingQueryMode),
+				connector.WithQueryBind(bind.TablePathPrefix("path/to/tables")),
+				connector.WithQueryBind(bind.PositionalArgs{}),
 			},
 			err: nil,
 		},
@@ -128,10 +123,10 @@ func TestParse(t *testing.T) {
 				config.WithEndpoint("localhost:2135"),
 				config.WithDatabase("/local"),
 			},
-			connectorOpts: []xsql.ConnectorOption{
-				xsql.WithDefaultQueryMode(xsql.ScriptingQueryMode),
-				xsql.WithTablePathPrefix("path/to/tables"),
-				xsql.WithQueryBind(bind.AutoDeclare{}),
+			connectorOpts: []connector.Option{
+				connector.WithDefaultQueryMode(tableSql.ScriptingQueryMode),
+				connector.WithQueryBind(bind.TablePathPrefix("path/to/tables")),
+				connector.WithQueryBind(bind.AutoDeclare{}),
 			},
 			err: nil,
 		},
@@ -142,9 +137,9 @@ func TestParse(t *testing.T) {
 				config.WithEndpoint("localhost:2135"),
 				config.WithDatabase("/local"),
 			},
-			connectorOpts: []xsql.ConnectorOption{
-				xsql.WithDefaultQueryMode(xsql.ScriptingQueryMode),
-				xsql.WithTablePathPrefix("path/to/tables"),
+			connectorOpts: []connector.Option{
+				connector.WithDefaultQueryMode(tableSql.ScriptingQueryMode),
+				connector.WithQueryBind(bind.TablePathPrefix("path/to/tables")),
 			},
 			err: nil,
 		},
@@ -155,11 +150,24 @@ func TestParse(t *testing.T) {
 				config.WithEndpoint("localhost:2135"),
 				config.WithDatabase("/local"),
 			},
-			connectorOpts: []xsql.ConnectorOption{
-				xsql.WithDefaultQueryMode(xsql.ScriptingQueryMode),
-				xsql.WithTablePathPrefix("path/to/tables"),
-				xsql.WithQueryBind(bind.PositionalArgs{}),
-				xsql.WithQueryBind(bind.AutoDeclare{}),
+			connectorOpts: []connector.Option{
+				connector.WithDefaultQueryMode(tableSql.ScriptingQueryMode),
+				connector.WithQueryBind(bind.TablePathPrefix("path/to/tables")),
+				connector.WithQueryBind(bind.PositionalArgs{}),
+				connector.WithQueryBind(bind.AutoDeclare{}),
+			},
+			err: nil,
+		},
+		{
+			dsn: "grpc://localhost:2135/local?go_fake_tx=scripting,scheme",
+			opts: []config.Option{
+				config.WithSecure(false),
+				config.WithEndpoint("localhost:2135"),
+				config.WithDatabase("/local"),
+			},
+			connectorOpts: []connector.Option{
+				connector.WithFakeTx(tableSql.ScriptingQueryMode),
+				connector.WithFakeTx(tableSql.SchemeQueryMode),
 			},
 			err: nil,
 		},
@@ -172,7 +180,20 @@ func TestParse(t *testing.T) {
 				require.NoError(t, err)
 				d, err := newConnectionFromOptions(context.Background(), opts...)
 				require.NoError(t, err)
-				require.Equal(t, newConnector(tt.connectorOpts...), newConnector(d.databaseSQLOptions...))
+				exp := newConnector(tt.connectorOpts...)
+				act := newConnector(d.databaseSQLOptions...)
+				t.Run("tableOptions", func(t *testing.T) {
+					require.Equal(t, newTableConn(exp.TableOpts...), newTableConn(act.TableOpts...))
+				})
+				t.Run("queryOptions", func(t *testing.T) {
+					require.Equal(t, newQueryConn(exp.QueryOpts...), newQueryConn(act.QueryOpts...))
+				})
+				exp.TableOpts = nil
+				exp.QueryOpts = nil
+				act.TableOpts = nil
+				act.QueryOpts = nil
+				require.Equal(t, exp.Bindings(), act.Bindings())
+				require.Equal(t, exp, act)
 				compareConfigs(t, config.New(tt.opts...), d.config)
 			}
 		})