+++ internal/connector/errors.go
@@ -1,0 +1,12 @@
+package connector
+
+import (
+	"database/sql/driver"
+	"errors"
+)
+
+var (
+	ErrUnsupported         = driver.ErrSkip
+	errAlreadyClosed       = errors.New("already closed")
+	errWrongQueryProcessor = errors.New("wrong query processor")
+)