+++ internal/query/conn/errors.go
@@ -1,0 +1,5 @@
+package conn
+
+import "errors"
+
+var errConnClosedEarly = errors.New("Conn closed early")