+++ internal/table/conn/stmt.go
@@ -1,4 +1,4 @@
-package xsql
+package conn
 
 import (
 	"context"
@@ -6,21 +6,19 @@ import (
 	"fmt"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/stack"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/table/conn/badconn"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/badconn"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
 type stmt struct {
-	conn      *conn
+	conn      *Conn
 	processor interface {
 		driver.ExecerContext
 		driver.QueryerContext
 	}
 	query string
 	ctx   context.Context //nolint:containedctx
-
-	trace *trace.DatabaseSQL
 }
 
 var (
@@ -30,8 +28,8 @@ var (
 )
 
 func (stmt *stmt) QueryContext(ctx context.Context, args []driver.NamedValue) (_ driver.Rows, finalErr error) {
-	onDone := trace.DatabaseSQLOnStmtQuery(stmt.trace, &ctx,
-		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql.(*stmt).QueryContext"),
+	onDone := trace.DatabaseSQLOnStmtQuery(stmt.conn.parent.Trace(), &ctx,
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table/conn.(*stmt).QueryContext"),
 		stmt.ctx, stmt.query,
 	)
 	defer func() {
@@ -49,8 +47,8 @@ func (stmt *stmt) QueryContext(ctx context.Context, args []driver.NamedValue) (_
 }
 
 func (stmt *stmt) ExecContext(ctx context.Context, args []driver.NamedValue) (_ driver.Result, finalErr error) {
-	onDone := trace.DatabaseSQLOnStmtExec(stmt.trace, &ctx,
-		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql.(*stmt).ExecContext"),
+	onDone := trace.DatabaseSQLOnStmtExec(stmt.conn.parent.Trace(), &ctx,
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table/conn.(*stmt).ExecContext"),
 		stmt.ctx, stmt.query,
 	)
 	defer func() {
@@ -74,8 +72,8 @@ func (stmt *stmt) NumInput() int {
 func (stmt *stmt) Close() (finalErr error) {
 	var (
 		ctx    = stmt.ctx
-		onDone = trace.DatabaseSQLOnStmtClose(stmt.trace, &ctx,
-			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql.(*stmt).Close"),
+		onDone = trace.DatabaseSQLOnStmtClose(stmt.conn.parent.Trace(), &ctx,
+			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table/conn.(*stmt).Close"),
 		)
 	)
 	defer func() {