+++ internal/table/conn/tx_fake.go
@@ -1,13 +1,13 @@
-package xsql
+package conn
 
 import (
 	"context"
 	"database/sql/driver"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/stack"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/table/conn/badconn"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/tx"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/badconn"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
@@ -15,13 +15,13 @@ type txFake struct {
 	tx.Identifier
 
 	beginCtx context.Context //nolint:containedctx
-	conn     *conn
+	conn     *Conn
 	ctx      context.Context //nolint:containedctx
 }
 
 func (tx *txFake) PrepareContext(ctx context.Context, query string) (_ driver.Stmt, finalErr error) {
-	onDone := trace.DatabaseSQLOnTxPrepare(tx.conn.trace, &ctx,
-		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql.(*txFake).PrepareContext"),
+	onDone := trace.DatabaseSQLOnTxPrepare(tx.conn.parent.Trace(), &ctx,
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table/conn.(*txFake).PrepareContext"),
 		tx.beginCtx, tx, query,
 	)
 	defer func() {
@@ -36,7 +36,6 @@ func (tx *txFake) PrepareContext(ctx context.Context, query string) (_ driver.St
 		processor: tx,
 		ctx:       ctx,
 		query:     query,
-		trace:     tx.conn.trace,
 	}, nil
 }
 
@@ -47,19 +46,19 @@ var (
 	_ tx.Identifier         = &txFake{}
 )
 
-func (c *conn) beginTxFake(ctx context.Context, txOptions driver.TxOptions) (currentTx, error) {
+func beginTxFake(ctx context.Context, c *Conn) currentTx {
 	return &txFake{
 		Identifier: tx.ID("FAKE"),
 		conn:       c,
 		ctx:        ctx,
-	}, nil
+	}
 }
 
 func (tx *txFake) Commit() (err error) {
 	var (
 		ctx    = tx.ctx
-		onDone = trace.DatabaseSQLOnTxCommit(tx.conn.trace, &ctx,
-			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql.(*txFake).Commit"),
+		onDone = trace.DatabaseSQLOnTxCommit(tx.conn.parent.Trace(), &ctx,
+			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table/conn.(*txFake).Commit"),
 			tx,
 		)
 	)
@@ -79,8 +78,8 @@ func (tx *txFake) Commit() (err error) {
 func (tx *txFake) Rollback() (err error) {
 	var (
 		ctx    = tx.ctx
-		onDone = trace.DatabaseSQLOnTxRollback(tx.conn.trace, &ctx,
-			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql.(*txFake).Rollback"),
+		onDone = trace.DatabaseSQLOnTxRollback(tx.conn.parent.Trace(), &ctx,
+			stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table/conn.(*txFake).Rollback"),
 			tx,
 		)
 	)
@@ -100,9 +99,8 @@ func (tx *txFake) Rollback() (err error) {
 func (tx *txFake) QueryContext(ctx context.Context, query string, args []driver.NamedValue) (
 	rows driver.Rows, err error,
 ) {
-	onDone := trace.DatabaseSQLOnTxQuery(
-		tx.conn.trace, &ctx,
-		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql.(*txFake).QueryContext"),
+	onDone := trace.DatabaseSQLOnTxQuery(tx.conn.parent.Trace(), &ctx,
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table/conn.(*txFake).QueryContext"),
 		tx.ctx, tx, query,
 	)
 	defer func() {
@@ -119,9 +117,8 @@ func (tx *txFake) QueryContext(ctx context.Context, query string, args []driver.
 func (tx *txFake) ExecContext(ctx context.Context, query string, args []driver.NamedValue) (
 	result driver.Result, err error,
 ) {
-	onDone := trace.DatabaseSQLOnTxExec(
-		tx.conn.trace, &ctx,
-		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql.(*txFake).ExecContext"),
+	onDone := trace.DatabaseSQLOnTxExec(tx.conn.parent.Trace(), &ctx,
+		stack.FunctionID("github.com/ydb-platform/ydb-go-sdk/v3/internal/table/conn.(*txFake).ExecContext"),
 		tx.ctx, tx, query,
 	)
 	defer func() {