+++ internal/table/conn/valuer.go
@@ -1,4 +1,4 @@
-package xsql
+package conn
 
 import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/scanner"